Build scripts updated

+++ build_master.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
+
 export PLATFORM="TW"
-export MREV="KK4.4"
 export CURDATE=`date "+%m.%d.%Y"`
 export MUXEDNAMELONG="KT-SGS4-$MREV-$PLATFORM-$CARRIER-$CURDATE"
 export MUXEDNAMESHRT="KT-SGS4-$MREV-$PLATFORM-$CARRIER*"
@@ -8,54 +8,66 @@ export KTVER="--$MUXEDNAMELONG--"
 export KERNELDIR=`readlink -f .`
 export PARENT_DIR=`readlink -f ..`
 export INITRAMFS_DEST=$KERNELDIR/kernel/usr/initramfs
-export INITRAMFS_SOURCE=`readlink -f ..`/Ramdisks/$PLATFORM"_"$CARRIER"4.4"
+export INITRAMFS_SOURCE=`readlink -f ..`/Ramdisk
+export INITRAMFS_BRANCH=$(echo $PLATFORM | awk '{print tolower($0)}')"-"$VERSION
 export CONFIG_$PLATFORM_BUILD=y
 export PACKAGEDIR=$PARENT_DIR/Packages/$PLATFORM
+
 #Enable FIPS mode
 export USE_SEC_FIPS_MODE=true
 export ARCH=arm
-# export CROSS_COMPILE=/home/ktoonsez/aokp4.2/prebuilts/gcc/linux-x86/arm/arm-eabi-4.6/bin/arm-eabi-
-#export CROSS_COMPILE=$PARENT_DIR/linaro4.9-a15/bin/arm-cortex_a15-linux-gnueabihf-
-export CROSS_COMPILE=/media/storage/toolchain/linaro-4.7-12.10/bin/arm-linux-gnueabihf-
-#export CROSS_COMPILE=/media/storage/CM11/prebuilts/gcc/linux-x86/arm/arm-eabi-4.7/bin/arm-eabi-
 
+# export CROSS_COMPILE=/home/ktoonsez/aokp4.2/prebuilts/gcc/linux-x86/arm/arm-eabi-4.6/bin/arm-eabi-
+#export CROSS_COMPILE=$PARENT_DIR/linaro4.7/bin/arm-eabi-
+export CROSS_COMPILE=/media/storage/toolchain/linaro-4.7-12.10/bin/arm-linux-gnueabihf-
+#export CROSS_COMPILE=/media/storage/toolchain/arm-linux-androideabi-4.8/bin/arm-linux-androideabi-
+#export CROSS_COMPILE=/media/storage/toolchain/sabermod-arm-linux-androideabi-4.9/bin/arm-linux-androideabi-
+#export CROSS_COMPILE=/media/storage/toolchain/gcc-linaro-arm-linux-gnueabihf-4.9-2014.09_linux/bin/arm-linux-gnueabihf-
+
+echo "** Checkout initramfs"
+cd $INITRAMFS_SOURCE
+git checkout $INITRAMFS_BRANCH
+if [ ! $? -eq 0 ]; then
+  exit 1
+fi
+cd $KERNELDIR
 
 time_start=$(date +%s.%N)
 
-echo "Remove old Package Files"
+echo "** Remove old Package Files"
 rm -rf $PACKAGEDIR/*
 
-echo "Setup Package Directory"
+echo "** Setup Package Directory"
 mkdir -p $PACKAGEDIR/system/app
 mkdir -p $PACKAGEDIR/system/lib/modules
 mkdir -p $PACKAGEDIR/system/etc/init.d
 
-echo "Create initramfs dir"
+echo "** Create initramfs dir"
 mkdir -p $INITRAMFS_DEST
 
-echo "Remove old initramfs dir"
+echo "** Remove old initramfs dir"
 rm -rf $INITRAMFS_DEST/*
 
-echo "Copy new initramfs dir"
+echo "** Copy new initramfs dir"
 cp -R $INITRAMFS_SOURCE/* $INITRAMFS_DEST
 
-echo "chmod initramfs dir"
+echo "** chmod initramfs dir"
 chmod -R g-w $INITRAMFS_DEST/*
 rm $(find $INITRAMFS_DEST -name EMPTY_DIRECTORY -print)
 rm -rf $(find $INITRAMFS_DEST -name .git -print)
 
-echo "Remove old zImage"
+echo "** Remove old zImage"
 rm $PACKAGEDIR/zImage
 rm arch/arm/boot/zImage
 
-echo "$BOARD"
+echo "** Board: $BOARD"
 if [ -z $BOARD ]; then
     export BOARD="jf"
 fi
 
-echo "Make the kernel"
+echo "** Make the kernel"
 make VARIANT_DEFCONFIG=${BOARD}_${CARRIER}"_defconfig" KT_jf_defconfig SELINUX_DEFCONFIG=selinux_defconfig
-echo "Modding .config file - "$KTVER
+echo "** Modding .config file - "$KTVER
 sed -i 's,CONFIG_LOCALVERSION="-KT-SGS4",CONFIG_LOCALVERSION="'$KTVER'",' .config
 
 HOST_CHECK=`uname -n`
@@ -67,7 +79,7 @@ else
 	make -j`grep 'processor' /proc/cpuinfo | wc -l`
 fi;
 
-echo "Copy modules to Package"
+echo "** Copy modules to Package"
 cp -a $(find . -name *.ko -print |grep -v initramfs) $PACKAGEDIR/system/lib/modules/
 if [ $ADD_KTWEAKER = 'Y' ]; then
 	cp $PARENT_DIR/ktapps/com.ktoonsez.KTweaker.apk $PACKAGEDIR/system/app/com.ktoonsez.KTweaker.apk
@@ -75,37 +87,39 @@ if [ $ADD_KTWEAKER = 'Y' ]; then
 fi;
 
 if [ -e $KERNELDIR/arch/arm/boot/zImage ]; then
-	echo "Copy zImage to Package"
+	echo "** Copy zImage to Package"
 	cp arch/arm/boot/zImage $PACKAGEDIR/zImage
 
-	echo "Make boot.img"
+	echo "** Make boot.img"
 	./mkbootfs $INITRAMFS_DEST | gzip > $PACKAGEDIR/ramdisk.gz
-	./mkbootimg --cmdline 'console=null androidboot.hardware=qcom user_debug=31 msm_rtb.filter=0x3F ehci-hcd.park=3 maxcpus=4' --kernel $PACKAGEDIR/zImage --ramdisk $PACKAGEDIR/ramdisk.gz --base 0x80200000 --pagesize 2048 --ramdisk_offset 0x02000000 --output $PACKAGEDIR/boot.img 
+	./mkbootimg --cmdline "$RD_CMDLINE" --kernel $PACKAGEDIR/zImage --ramdisk $PACKAGEDIR/ramdisk.gz --base 0x80200000 --pagesize 2048 --ramdisk_offset 0x02000000 --output $PACKAGEDIR/boot.img 
 	#if [ $EXEC_LOKI = 'Y' ]; then
 	#	echo "Executing loki"
 	#	./loki_patch-linux-x86_64 boot aboot$CARRIER.img $PACKAGEDIR/boot.img $PACKAGEDIR/boot.lok
 	#	rm $PACKAGEDIR/boot.img
 	#fi;
 	cd $PACKAGEDIR
+
 	#if [ $EXEC_LOKI = 'Y' ]; then
 	#	cp -R ../META-INF-SEC ./META-INF
 	#else
 		cp -R $PARENT_DIR/ktapps/META-INF .
 	#fi;
 	cp -R ../kernel .
+
 	rm ramdisk.gz
 	rm zImage
 	rm ../$MUXEDNAMESHRT.zip
 	zip -r ../$MUXEDNAMELONG.zip .
 
 	time_end=$(date +%s.%N)
-	echo -e "${BLDYLW}Total time elapsed: ${TCTCLR}${TXTGRN}$(echo "($time_end - $time_start) / 60"|bc ) ${TXTYLW}minutes${TXTGRN} ($(echo "$time_end - $time_start"|bc ) ${TXTYLW}seconds) ${TXTCLR}"
+	echo -e "** ${BLDYLW}Total time elapsed: ${TCTCLR}${TXTGRN}$(echo "($time_end - $time_start) / 60"|bc ) ${TXTYLW}minutes${TXTGRN} ($(echo "$time_end - $time_start"|bc ) ${TXTYLW}seconds) ${TXTCLR}"
 
-	export DLNAME="http://ktoonsez.jonathanjsimon.com/sgs4/$PLATFORM/$MUXEDNAMELONG.zip"
+	#export DLNAME="http://ktoonsez.jonathanjsimon.com/sgs4/$PLATFORM/$MUXEDNAMELONG.zip"
 	
-	FILENAME=../$MUXEDNAMELONG.zip
-	FILESIZE=$(stat -c%s "$FILENAME")
-	echo "Size of $FILENAME = $FILESIZE bytes."
+	#FILENAME=../$MUXEDNAMELONG.zip
+	#FILESIZE=$(stat -c%s "$FILENAME")
+	#echo "Size of $FILENAME = $FILESIZE bytes."
 	#rm ../$MREV-$PLATFORM-$CARRIER"-version.txt"
 	#exec 1>>../$MREV-$PLATFORM-$CARRIER"-version.txt" 2>&1
 	#echo -n "$MUXEDNAMELONG,$FILESIZE," & curl -s https://www.googleapis.com/urlshortener/v1/url --header 'Content-Type: application/json' --data "{'longUrl': '$DLNAME'}" | grep \"id\" | sed -e 's,^.*id": ",,' -e 's/",.*$//'
@@ -122,5 +136,5 @@ if [ -e $KERNELDIR/arch/arm/boot/zImage ]; then
 
 	cd $KERNELDIR
 else
-	echo "KERNEL DID NOT BUILD! no zImage exist"
+	echo "** KERNEL DID NOT BUILD! no zImage exist"
 fi;
